Audible click feedback when saving position
The click sound is implemented as a PRE_MOVE, single DOWN, PRE_STOP
command sequence. This doesn't actually move the desk, but causes some
hardware in the legs (relays?) to click on and off as they normally do
before and after moving.

Only sending PRE_MOVE and PRE_STOP was not enough to cause a click. So
it seems the single movement command is necessary. Choosing DOWN vs UP
was arbitrary.